backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
PyMuPDF
pytesseract
Pillow
numpy
chromadb
faiss-cpu
langchain
llama-index
haystack-ai
httpx
python-dotenv
jsonschema
tenacity
loguru
google-generativeai
reportlab
pytest
sqlmodel
psycopg[binary]
